// newcode/libs/ui/common/src/lib/chart/entity/chart-type.enum.ts
/*
 * $Id$
 * Copyright(c) RIB Software GmbH
 */


/**
 *  Chart type enum
 */
export enum ChartTypeEnum {
	bar = 'bar',
	horizontalBar = 'horizontalBar',
	line = 'line',
	radar = 'radar',
	pie = 'pie',
	bubble = 'bubble',
	doughnut = 'doughnut',
	polarArea = 'polarArea',
	scatter = 'scatter',
	/*
	 * todo: a extensional type, need to enhance
	 */
	threeD_Columns = '3d_Columns',
}